Squashed of recursive function invocation commits:
commit aa9c62503fb322928c516adbb9d8df5ae3cb5fb8
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sat Oct 29 23:15:13 2016 -0700

    Recursive invocation test

commit 071f190ba396fe8ba3486a3c946d97095335c2a9
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sat Oct 29 13:01:03 2016 -0700

    Final puzzle pieces for Maybe type matching

commit 44f8ba8aba8e547df4aa5731e1e99777aa859e6e
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sat Oct 29 12:04:18 2016 -0700

    Re-purpose tests that were shown to have semantic correctness issues

commit fd0793646796347322ac0efe7583eeeb9894ee22
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Oct 25 22:29:40 2016 -0700

    Checkpoint commit on the road to Maybe types as sum type subclass

commit b4ae70b20b90cbaaa442c7115d3f831d55e99f15
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Oct 25 21:53:27 2016 -0700

    Constructor/destructor housekeeping

commit 38ac51873e762d15043cc9800e091627e862421b
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Fri Oct 7 23:22:49 2016 -0700

    Enforce lexicographic order test ordering

commit 29b0c01db840a085f12371f68d04437759f2be62
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Wed Sep 21 18:53:38 2016 -0700

    Maybe type assignment fixes

commit 3daaff27b12bb4427dc253a27224d958eec84e7a
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Sep 20 22:53:29 2016 -0700

    Test invalid passing of maybe type as argument

commit 0221b6998dda74dfc67a5b86292a981bfd3de763
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Sep 20 22:28:35 2016 -0700

    Fix runtime error reporting logic during function evaluation

commit da95996ed4dfc8479b0824c087b7083f8701e6e9
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Sep 13 23:04:07 2016 -0700

    Baby's first linked list processor

commit 6b052c58eeb9fa2e0ffa5434a0815710ab9f3de0
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Sep 13 22:56:42 2016 -0700

    Refine and generalize argument value assignment logic

commit ead1f206a4ab99242804aaca1f5be130d86de6c7
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Sep 11 21:53:05 2016 -0700

    TResult -> TypedResult

commit 70cb27974385f7d9299c0c5d1c21541acd3419a8
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Sep 11 21:38:25 2016 -0700

    Return TResult<string> from Expression::ToString

commit 8755cc4bac9305360057bd25529e1b5aa2715862
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Sep 11 21:10:58 2016 -0700

    Test duplicate recursive function definition

commit 223ff4f5930389329f74fe0421a19468f066b9fd
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Sep 11 21:09:30 2016 -0700

    Housekeeping

commit 4292f8203d71294ce5b93cc2d91e7ab48dc594fb
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sat Sep 10 10:04:13 2016 -0700

    Support for type-inferred recursive functions

    Supporting changes:

    * Return Result from Expression::GetTypeSpecifier and
    Variable::GetTypeSpecifier
    * Excise GetReturnStatementErrors API in favor of expanded Preprocess
    interface

commit 967083291d0d7f31ae470cb79e9ff0540f2570f9
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Thu Aug 11 20:20:59 2016 -0700

    Minimal recursive function support

commit 46dc4f89b321b0538cc6c2e1b411b418da604c48
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Jul 31 23:14:02 2016 -0700

    Fixes for issues surfaced by upgrading built environment

    New environment:
    * GCC 5.4.0
    * Eclipse Neon

commit 8f083abba67d86a82d45f203863db7b655f20021
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sat Jul 23 22:44:39 2016 -0700

    Add and utilize a closure context during preprocessing and execution.

    This addition was made to fix memory leak in tests t3030, t3031, etc. In
    these tests, strong references to the temporary member-generation contexts were established by the function closure-referencing logic. This led to a reference loop in the temporary context.

    Adding an output context instead of a closure context was also considered,
    but the overhead of establishing a complex type's validity in the output context seemed too high.

commit b6fa640ac7c36d7e3b16c764b24af07da527234e
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Jul 5 16:04:58 2016 -0700

    Allow variable declarations in match contexts

commit 43b98f9e8f421f19a9dd28512eb20149c8221eb8
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Tue Jul 5 14:50:22 2016 -0700

    Naming clarification

commit 7e148712778fa868366522eecb78b59a3d39b73c
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sat Jul 2 16:36:46 2016 -0700

    Prevent symbol leakage from for and if block contexts

commit f3717f0de752b2b582745387c2dd4ca9c71298e6
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Jun 19 22:57:03 2016 -0700

    More complicated function-as-record-member testing

commit b72c76018b45a05fdacfdb5a5d1fcc95f97fb192
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Jun 19 22:56:10 2016 -0700

    Naming normalization

commit 5026c063c03837c8317dcca643f4b286532199f5
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Jun 19 18:21:56 2016 -0700

    Ignore generated gcov files

commit 8871f680a7896ffea25726db41bbd080f913fedf
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Jun 19 18:17:10 2016 -0700

    Cleaner cleaning process in Makefile

commit 2507708a9b607f85f3afcc8ab000842c153a8054
Author: Caleb Callaway <cqcallaw@gmail.com>
Date:   Sun Jun 19 17:27:23 2016 -0700

    Fix linting issues

+++ src/array.h
@@ -62,11 +62,11 @@ public:
 		if (index < GetSize()) {
 			new_vector->at(index) = value;
 		} else {
-			int old_size = new_vector->size();
+			size_t old_size = new_vector->size();
 			new_vector->resize(index);
 
 			//fill with default values
-			for (int i = old_size; i < new_vector->size(); i++) {
+			for (size_t i = old_size; i < new_vector->size(); i++) {
 				const_shared_ptr<void> default_value =
 						GetElementTypeSpecifier()->DefaultValue(type_table);
 				new_vector->at(i) = default_value;
